Add ability to wait/retry until changes show up
